Исправлен timer
