+++ package.json
@@ -2,7 +2,7 @@
   "name": "test-ssigment",
   "version": "0.1.0",
   "private": true,
-  "homepage": "http://rpopov94.github.io/test-ssigment/",
+  "homepage": "https://rpopov94.github.io/test-ssigment",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
